packages/response-parser/src/response/response.interceptor.ts
import { Injectable, NestInterceptor, ExecutionContext, CallHandler } from '@nestjs/common';
import { Observable } from 'rxjs';
import { map } from 'rxjs/operators';
import * as _ from 'lodash';

@Injectable()
export class ResponseInterceptor implements NestInterceptor {
  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
    return next.handle().pipe(
      map((res) => {
        let parse = { data: [] };
        if (_.isArray(res)) {
          parse.data = res;
        } else {
          parse = res;
        }
        return parse;
      }),
    );
  }
}
